requirements.txt
django~=3.1.6
django-choices
psycopg2
Pillow
python-dotenv~=0.15.0
djangorestframework~=3.12.4
djangorestframework_simplejwt
djangorestframework-jwt
pip~=21.0.1
configparser~=5.0.2
socks
cryptography~=3.4.7
py~=1.10.0
lxml~=4.6.3
setuptools~=49.2.1
factory_boy
Faker~=6.6.3
django-cors-headers==3.7.0
celery~=5.0.0
redis~=3.5.3
django-redis~=4.12.1
gunicorn==20.0.4
black~=21.5b1
pre-commit~=2.12.1
drf-yasg~=1.20.0
djangorestframework-jwt
